requirements.txt
boto3
tqdm
python-dotenv
pinecone-client
transformers
torch
pandas
nltk
spacy
yake
pyyaml
numpy